feat(serializer): add validation fields to ResponseSerializationError
Mirror the request-side error shape by exposing `validation:
FastifySchemaValidationError[]` and `validationContext = 'body' as const`
on ResponseSerializationError. Each entry's schemaPath is prefixed
`#/body/<path>`, so shared error handlers can now treat request and
response validation errors uniformly.

Relocate `mapIssueToValidationError` from `src/validator/error.ts` to
`src/utils/map-issue-to-validation-error.ts` — a neutral location
reachable from both the validator and serializer without cross-module
coupling. The public export from the package root is unchanged.
