+++ src/serializer/error.ts
@@ -1,11 +1,15 @@
+import type { FastifySchemaValidationError } from 'fastify';
 import type z from 'zod';
+
+import { mapIssueToValidationError } from '../utils/map-issue-to-validation-error.js';
 
 /**
  * Error thrown when response serialization fails.
  *
  * Thrown by `serializerCompiler` and `parseSerializerCompiler` when the handler's
  * return value does not match the response schema. Contains the HTTP method, URL,
- * and the underlying `ZodError` for inspection.
+ * the underlying `ZodError`, and a Fastify-compatible `validation` array with
+ * `schemaPath` entries prefixed `#/body/...` — mirroring `RequestValidationError`.
  *
  * Use `instanceof` to catch in a Fastify error handler. The `code` property
  * (`'ERR_RESPONSE_SERIALIZATION'`) is stable for programmatic matching.
@@ -16,7 +20,13 @@ import type z from 'zod';
  *   if (error instanceof ResponseSerializationError) {
  *     // Do NOT forward error.message or error.zodError to the client —
  *     // they contain internal schema details. Log them server-side instead.
- *     request.log.error({ method: error.method, url: error.url, httpStatus: error.httpStatus, zodError: error.zodError });
+ *     request.log.error({
+ *       method: error.method,
+ *       url: error.url,
+ *       httpStatus: error.httpStatus,
+ *       validation: error.validation,       // FastifySchemaValidationError[]
+ *       context: error.validationContext,   // always 'body'
+ *     });
  *     reply.code(500).send({
  *       error: error.code,         // 'ERR_RESPONSE_SERIALIZATION'
  *       method: error.method,      // 'GET'
@@ -34,6 +44,8 @@ export class ResponseSerializationError extends Error {
   readonly url: string;
   readonly httpStatus: string | undefined;
   readonly zodError: z.ZodError;
+  readonly validation: FastifySchemaValidationError[];
+  readonly validationContext = 'body' as const;
 
   constructor(
     options: { method: string; url: string; httpStatus?: string; zodError: z.ZodError },
@@ -48,5 +60,8 @@ export class ResponseSerializationError extends Error {
     this.url = options.url;
     this.httpStatus = options.httpStatus;
     this.zodError = options.zodError;
+    this.validation = options.zodError.issues.map((issue) =>
+      mapIssueToValidationError(issue, 'body'),
+    );
   }
 }
